Move connection-related commands from BaseRedisCommands to RedisConnectionCommands #2031
