+++ src/main/java/io/lettuce/core/api/sync/RedisServerCommands.java
@@ -20,10 +20,7 @@ import java.util.List;
 import java.util.Map;
 
 import io.lettuce.core.FlushMode;
-import io.lettuce.core.KillArgs;
 import io.lettuce.core.ShutdownArgs;
-import io.lettuce.core.TrackingArgs;
-import io.lettuce.core.UnblockType;
 import io.lettuce.core.protocol.CommandType;
 
 /**
@@ -50,99 +47,6 @@ public interface RedisServerCommands<K, V> {
      * @return String simple-string-reply.
      */
     String bgsave();
-
-    /**
-     * Control tracking of keys in the context of server-assisted client cache invalidation.
-     *
-     * @param enabled {@code true} to enable key tracking.
-     * @return String simple-string-reply {@code OK}.
-     * @since 6.0
-     */
-    String clientCaching(boolean enabled);
-
-    /**
-     * Get the current connection name.
-     *
-     * @return K bulk-string-reply The connection name, or a null bulk reply if no name is set.
-     */
-    K clientGetname();
-
-    /**
-     * Returns the client ID we are redirecting our tracking notifications to.
-     *
-     * @return the ID of the client we are redirecting the notifications to. The command returns -1 if client tracking is not
-     *         enabled, or 0 if client tracking is enabled but we are not redirecting the notifications to any client.
-     * @since 6.0
-     */
-    Long clientGetredir();
-
-    /**
-     * Get the id of the current connection.
-     *
-     * @return Long The command just returns the ID of the current connection.
-     * @since 5.3
-     */
-    Long clientId();
-
-    /**
-     * Kill the connection of a client identified by ip:port.
-     *
-     * @param addr ip:port.
-     * @return String simple-string-reply {@code OK} if the connection exists and has been closed.
-     */
-    String clientKill(String addr);
-
-    /**
-     * Kill connections of clients which are filtered by {@code killArgs}.
-     *
-     * @param killArgs args for the kill operation.
-     * @return Long integer-reply number of killed connections.
-     */
-    Long clientKill(KillArgs killArgs);
-
-    /**
-     * Get the list of client connections.
-     *
-     * @return String bulk-string-reply a unique string, formatted as follows: One client connection per line (separated by LF),
-     *         each line is composed of a succession of property=value fields separated by a space character.
-     */
-    String clientList();
-
-    /**
-     * Stop processing commands from clients for some time.
-     *
-     * @param timeout the timeout value in milliseconds.
-     * @return String simple-string-reply The command returns OK or an error if the timeout is invalid.
-     */
-    String clientPause(long timeout);
-
-    /**
-     * Set the current connection name.
-     *
-     * @param name the client name.
-     * @return simple-string-reply {@code OK} if the connection name was successfully set.
-     */
-    String clientSetname(K name);
-
-    /**
-     * Enables the tracking feature of the Redis server, that is used for server assisted client side caching. Tracking messages
-     * are either available when using the RESP3 protocol or through Pub/Sub notification when using RESP2.
-     *
-     * @param args for the CLIENT TRACKING operation.
-     * @return String simple-string-reply {@code OK}.
-     * @since 6.0
-     */
-    String clientTracking(TrackingArgs args);
-
-    /**
-     * Unblock the specified blocked client.
-     *
-     * @param id the client id.
-     * @param type unblock type.
-     * @return Long integer-reply number of unblocked connections.
-     * @since 5.1
-     */
-    Long clientUnblock(long id, UnblockType type);
 
     /**
      * Returns an array reply of details about all Redis commands.
@@ -420,6 +324,7 @@ public interface RedisServerCommands<K, V> {
      * @return String simple-string-reply.
      * @deprecated since 6.1.7, use {@link #replicaof(String, int)} instead.
      */
+    @Deprecated
     String slaveof(String host, int port);
 
     /**
@@ -428,6 +333,7 @@ public interface RedisServerCommands<K, V> {
      * @return String simple-string-reply.
      * @deprecated since 6.1.7, use {@link #replicaofNoOne()} instead.
      */
+    @Deprecated
     String slaveofNoOne();
 
     /**
