+++ src/main/java/io/lettuce/core/cluster/api/sync/NodeSelectionCommands.java
@@ -23,7 +23,8 @@ import io.lettuce.core.cluster.api.NodeSelectionSupport;
  * @author Mark Paluch
  */
 public interface NodeSelectionCommands<K, V>
-        extends BaseNodeSelectionCommands<K, V>, NodeSelectionGeoCommands<K, V>, NodeSelectionHashCommands<K, V>,
+        extends BaseNodeSelectionCommands<K, V>, NodeSelectionConnectionCommands<K, V>, NodeSelectionGeoCommands<K, V>,
+        NodeSelectionHashCommands<K, V>,
         NodeSelectionHLLCommands<K, V>, NodeSelectionKeyCommands<K, V>, NodeSelectionListCommands<K, V>,
         NodeSelectionScriptingCommands<K, V>, NodeSelectionServerCommands<K, V>, NodeSelectionSetCommands<K, V>,
         NodeSelectionSortedSetCommands<K, V>, NodeSelectionStreamCommands<K, V>, NodeSelectionStringCommands<K, V> {
